detached Clock subcomponents to separate folder
