--- src/components/ClockCell.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const ClockCell = ({ type, children }) => {
-	return <div className="clock-cell">{children}</div>;
-};
-
-export default ClockCell;
