+++ src/components/clock/Clock.js
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react';
-import { useTime } from '../hooks/useTime';
+import { useTime } from '../../hooks/useTime';
 
 import ClockHand from './ClockHand';
